fix: move_note takes f64 position per spec, remove bridging i32 cast
Change pub fn move_note signature from new_position: i32 to f64,
matching spec §6.3 (MoveNote.new_position: f64). Remove the as i32
bridging cast in apply_retract_inverse_internal and the as f64 cast
when building the Operation::MoveNote variant. Update all call sites
including the Tauri command and test helpers.

+++ krillnotes-desktop/src-tauri/src/lib.rs
@@ -1034,7 +1034,7 @@ fn move_note(
     state: State<'_, AppState>,
     note_id: String,
     new_parent_id: Option<String>,
-    new_position: i32,
+    new_position: f64,
 ) -> std::result::Result<(), String> {
     let label = window.label();
     let mut workspaces = state.workspaces.lock()
